Add files via upload
The rDNAmine_collection_inventory.sh program compares each module to the reference and assesses its quality. When running the program, it is essential to define -r and provide the path to the reference module. This program prepares module-reference comparison files, which will later be used for comprehensive analyses of module similarity. Our goal was to eliminate the need for global alignment of repeats, a process that is highly time-consuming and computationally demanding. Instead, each module is compared individually against the reference, and the set of such comparisons is transformed into easy-to-use tabular data at later stages of analysis.
Input:

    read_name.modul.fasta
    read_name.modul.r.fasta

Output:

    read_name.modul.pair.fasta: Contains the reference and module aligned using the MAFFT program.
    read_name.modul.pair.quality: Contains the filename of the alignment, the number of gaps ("-") in the reference sequence, and the number of gaps in the module.
    read_name.modul.pair.fasta: Contains the sequences of the reference and module.

Output in rDNA_moduls_mapped directory:

Files for traditional analysis of all modules are stored in the rDNA_modules_mapped directory:

    modul.fasta: FASTA file containing all modules and the reference.
    modul.fastq: FASTQ file containing all modules and the reference.
    modul.sam: SAM file generated by mapping reads to the reference using Minimap2.
    modul.bam: BAM file generated by mapping reads to the reference using Minimap2.
    modul.sor.bam: Sorted BAM file produced from the mapped reads using Samtools. This format is viewable in IGV.
    modul_mapped.bam: BAM file containing only the modules mapped to the reference.
    all_rep_qq.csv: CSV file compiling data from all read_name.modul.pair.quality files.
    all_multi.fasta: FASTA file containing valid modules and the reference.
    read_names_modul_all.txt: List of all module names.