Add files via upload
This program extracts regions of reads identified as repeats and saves them into separate FASTA files. The original read name is preserved in the names of the resulting files. It is possible to define a length range for the extracted sequences. However, it is important to note that if the repeated module is 9,000 bp, setting the lower limit to, for example, 2,000 bp will result in the inclusion of significantly truncated modules in the subsequent analysis, which can complicate downstream processing. We recommend using a range of (module length - 10%; module length + 10%).

Input:

    FASTA files containing reads with identified repeats.
    Corresponding .tab.out files with coordinates of repeats on the reads.

Output:

    read_name.modul.fasta: FASTA file with the module sequence.
    read_name.modul.r.fasta: FASTA file with the module sequence in reverse conformation.
    read_name.tab.c.out: Tabular file with coordinates.

Output Organisation:

    rDNA_irrelevant: Contains all input files and read_name.tab.c.out files.
    rDNA_modules: Contains files with module sequences.