Add files via upload
The rDNAmine_prospector.sh program uses Markov chains to identify reads containing repeat sequences. It requires the number of threads available for computation and the path to a FASTA file containing the sequence of a single repeat as input arguments. If the full rDNA module sequence is known for a given organism, it should be provided as the reference. If not, a sequence from a closely related organism can be used.

Alternatively, a fragment of the module can be used as the reference. In such cases, the search must be conducted in two steps. First, coordinates of the full repeat present on the reads are determined. Then, by saving the fragment of the read containing the complete repeat to a FASTA file, a reference is created. This reference is then used to repeat the search.

Input:

    FASTA files containing individual reads.
    A directory containing these files, prepared by the rDNAmine_miner program, which adds the suffix .split to the filenames.

Output:

    read_name.mine.out files: Contain the output from the nhmmer program.
    read_name.tab.out files: Contain coordinates of identified repeats on the respective reads.

All intermediate analysis files are organised into directories:

    .mine.out files are stored in rDNA_repetitions/rDNA_mine.
    .tab.out files and FASTA files of reads without the specified repeats are placed in rDNA_repetitions/rDNA_pile.

In the working directory, only FASTA files of reads with identified repeats and .tab.out files with the coordinates of identified repeats remain.

The .tab.out files can be used to determine the lengths of the identified repeats and to check for any chimeric repeats—those with a structure different from the expected one.